fix: Project Restructure
