+++ .env.example
@@ -1,0 +1,2 @@
+PORT=3000
+JWT_SECRET=my_secret
