--- .env_example
@@ -1,2 +1,0 @@
-PORT=5000
-JWT_SECRET=YOUR_SECRET
