+++ .eslintrc.json
@@ -1,0 +1,15 @@
+{
+	"env": {
+		"es2021": true,
+		"node": true
+	},
+	"parserOptions": {
+		"ecmaVersion": "latest",
+		"sourceType": "module"
+	},
+	"rules": {
+		"linebreak-style": ["error", "unix"],
+		"quotes": ["error", "double"],
+		"semi":  ["error", "always"]
+	}
+}
