--- nodemon.json
@@ -1,7 +1,0 @@
-{
-    "watch": ["src"],
-    "ext": ".js",
-    "ignore": [],
-    "exec": "node src/server.js"
-  }
-
