+++ package.json
@@ -5,20 +5,25 @@
   "repository": "https://github.com/FrontendArmy/logoexecutive-backend.git",
   "license": "MIT",
   "scripts": {
-    "dev": "nodemon",
+    "dev": "NODE_ENV=development nodemon",
     "start": "node src/index.js",
+    "lint": "eslint .",
+    "lint:fix": "eslint --fix",
+    "lint:fix-all": "eslint --fix .",
     "test": "jest"
   },
   "dependencies": {
-    "body-parser": "^1.20.2",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "jsonwebtoken": "^9.0.2"
   },
   "devDependencies": {
+    "eslint": "^8.46.0",
     "jest": "^29.6.4",
     "nodemon": "^3.0.1",
+    "pre-commit": "^1.2.2",
     "supertest": "^6.3.3"
-  }
+  },
+	"pre-commit": ["lint"]
 }
