+++ src/db/users.js
@@ -1,6 +1,6 @@
 const users = [
   {
-    id: '1',
+    id: "1",
     username: "anonymous",
     password: "welcome123",
   },
