+++ src/middlewares/errorHandler.js
@@ -1,6 +1,6 @@
 const errorHandler = (err, _req, res, next) => {
   res.status(500).json({ success: false, message: err.message });
-  next()
+  next();
 };
 
-module.exports = errorHandler
+module.exports = errorHandler;
