+++ src/middlewares/routeNotFound.js
@@ -1,6 +1,6 @@
 const routeNotFound = (_req, res, next) => {
   res.status(404).json({ message: "route not found" });
-  next()
+  next();
 };
 
-module.exports = routeNotFound
+module.exports = routeNotFound;
