+++ src/routes/index.js
@@ -8,6 +8,6 @@ router.get("/", (_req, res) =>
   res.status(200).json({message: "Welcome to Logo executive API"})
 );
 
-router.use('/auth', auth)
+router.use("/auth", auth);
 
 module.exports = router;
